+++ src/GafferDispatch/TaskNode.cpp
@@ -56,11 +56,7 @@ using namespace GafferDispatch;
 //////////////////////////////////////////////////////////////////////////
 
 TaskNode::Task::Task( ConstTaskPlugPtr plug, const Gaffer::Context *context )
-	:	m_plug( plug ), m_context( new Context( *context ) )
-{
-}
-
-TaskNode::Task::Task( TaskNodePtr n, const Context *c ) : m_plug( n->taskPlug() ), m_context( new Context( *c ) )
+	:	m_plug( plug ), m_context( context )
 {
 }